add of all Test with jUnit

+++ src/main/java/com/mmn/circuitscourts/models/Article.java
@@ -102,7 +102,7 @@ public class Article {
         return numSiret;
     }
 
-    public void setNumSiret(int numSiret) {
+    public void setNumSiret(long numSiret) {
         this.numSiret = numSiret;
     }
